+++ apps/vite-demo/src/App.vue
@@ -1,12 +1,12 @@
 <script setup lang="ts">
 import HelloWorld from './components/HelloWorld.vue'
-import { add } from '@yolo/ui-utils'
+// import { add } from '@yolo/ui-utils'
 
-console.log(add(1,2))
+// console.log(add(1,2))
 </script>
 
 <template>
-  <pg-button>1221</pg-button>
+  <pg-button type="primary">1221</pg-button>
   <pg-tag>tag</pg-tag>
   <div>
     <a href="https://vite.dev" target="_blank">
