+++ apps/vite-demo/src/main.ts
@@ -2,6 +2,7 @@ import { createApp } from 'vue'
 import './style.css'
 import App from './App.vue'
 import PgKit from '@yolo/pg-kit'
+import "@yolo/pg-kit/dist/index.css"
 
 
 const app = createApp(App)
