--- packages/ui/pg-components/components/button/index.js
@@ -1,10 +1,0 @@
-import PgButton from './src/index.vue'
-
-const ButtonInstall = {
-  install(app) {
-    app.component('pg-button', PgButton);
-  }
-}
-
-export default ButtonInstall
-export { PgButton }
