+++ packages/ui/pg-components/components/button/index.ts
@@ -1,0 +1,11 @@
+import type { App, Plugin } from 'vue'
+import PgButton from './src/index.vue'
+
+const ButtonInstall: Plugin = {
+  install(app: App) {
+    app.component('pg-button', PgButton)
+  }
+}
+
+export default ButtonInstall
+export { PgButton }
