+++ packages/ui/pg-components/components/package.json
@@ -2,7 +2,7 @@
   "name": "@yolo/pg-components",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "index.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
