+++ packages/ui/pg-components/components/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../tsconfig.base.json",
+  "include": ["./*.ts", "./**/*.ts"]
+}
