+++ packages/ui/pg-components/ui/components.ts
@@ -1,4 +1,5 @@
 import ButtonInstall from '@yolo/pg-components/button'
 import TagInstall from '@yolo/pg-components/tag'
+import type { Plugin } from 'vue'
 
-export default [ButtonInstall, TagInstall]
+export default [ButtonInstall, TagInstall] as Plugin[]
