+++ packages/ui/pg-components/ui/package.json
@@ -12,10 +12,12 @@
   },
   "devDependencies": {
     "typescript": "~5.7.2",
+    "unocss": "^65.4.3",
     "vite": "^6.1.0"
   },
   "dependencies": {
     "@vitejs/plugin-vue": "^5.2.1",
-    "@yolo/pg-components": "workspace:^"
+    "@yolo/pg-components": "workspace:^",
+    "sass": "^1.84.0"
   }
 }
