+++ packages/ui/pg-components/ui/styles/button.scss
@@ -1,0 +1,8 @@
+.pg-button + .pg-button {
+  margin-left: 12px;
+}
+.pg-button {
+  border: none;
+  font-size: 14px;
+  color: red;
+}
