+++ packages/ui/pg-components/ui/styles/index.scss
@@ -1,0 +1,1 @@
+@use './button.scss';
