+++ packages/ui/pg-components/ui/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../tsconfig.base.json",
+  "include": ["./*.ts", "./**/*.ts"]
+}
